Move constants to Config class
