+++ src/dis_mc68000.h
@@ -32,9 +32,9 @@ public:
     DisOperand &getFormatter() override { return _formatter; }
     bool setCpu(const char *cpu) override { return TableMc68000.setCpu(cpu); }
     const char *listCpu() const override { return TableMc68000::listCpu(); }
-    Endian endian() const override { return ENDIAN_BIG; }
-    host::uint_t maxBytes() const override { return Entry::code_max; }
-    host::uint_t maxName() const override { return Entry::name_max; }
+    Endian endian() const override { return Config::endian; }
+    host::uint_t maxBytes() const override { return Config::code_max; }
+    host::uint_t maxName() const override { return Config::name_max; }
 
 private:
     DisMotoOperand _formatter;
