+++ src/entry_mc68000.h
@@ -66,8 +66,6 @@ struct Entry {
     const Config::insn_t insnCode;
     const host::uint_t flags;
     const char *name;
-    static constexpr host::uint_t code_max = 10;
-    static constexpr host::uint_t name_max = 7;
 
     static inline InsnFormat _insnFormat(host::uint_t flags) {
         return InsnFormat((flags >> insnFmt_gp) & insnFmt_gm);
