+++ src/insn_mc6800.h
@@ -23,7 +23,7 @@
 namespace libasm {
 namespace mc6800 {
 
-class InsnMc6800 : public InsnBase<ENDIAN_BIG, Config::uintptr_t> {
+class InsnMc6800 : public InsnBase<Config::endian, Config::uintptr_t> {
 public:
     InsnMc6800(Insn<Config::uintptr_t> &insn) : InsnBase(insn) {}
