+++ src/table_z80.cpp
@@ -333,7 +333,7 @@ static Error searchInsnCode(
         const Entry *entry = searchEntry(insn.opCode(), page->table, page->end);
         if (entry) {
             insn.setFlags(pgm_read_byte(&entry->flags1), pgm_read_byte(&entry->flags2));
-            char name[Entry::name_max + 1];
+            char name[Config::name_max + 1];
             pgm_strncpy(name, entry->name, sizeof(name));
             insn.setName(name);
             return OK;
